+++ src/main/java/com/ll/medium/domain/member/member/form/JoinForm.java
@@ -1,4 +1,4 @@
-package com.ll.medium.domain.member.member.dto;
+package com.ll.medium.domain.member.member.form;
 
 import jakarta.validation.constraints.NotBlank;
 import lombok.AllArgsConstructor;
@@ -6,7 +6,7 @@ import lombok.Getter;
 
 @Getter
 @AllArgsConstructor // 파라미터 받을 때 필수
-public class JoinDto {
+public class JoinForm {
     @NotBlank
     private String username;
     @NotBlank
